Add files via upload
The primary goals of this system are; [1] Make remote monitoring simple so you don't have to be close to the machine. [2] Accelerate the equipment's fault or potential hazard detection procedure. [3] Make data analysis and documentation easier for tracking performance and maintenance. [4] By providing early warning warnings, you can lower the chance of critical failures.
Key advantages; [1] Monitoring in real time with every two seconds, the data is refreshed. [2] Web access is available from any device with an internet connection. [3] IoT integration eliminates the requirement for a specialized computer that is always on site. [4] Scalable for suitable for a number of machines.